Stop narrowing wire values to Int, which traps on 32-bit
On armv7 Int is 32 bits and Swift traps -- not truncates -- on an
out-of-range conversion, with no compile-time diagnostic. Eight sites
narrowed unbounded values straight off the wire:

  Protobuf.parse        Int(key) & 7 and Int(key) >> 3 on a varint field key;
                        now masked and bounded by protobuf's 2^29-1 maximum
  Protobuf.readValue    .number(Int(value)) forced every 64-bit varint into a
                        platform Int; Value.number is now Int64
  Protobuf.readValue    slice(count: Int(length)) took an attacker-controlled
                        length prefix unbounded; now checked against the buffer
  UMPReader.readVarint  the 4-byte form and the shifted tail both exceed Int32;
                        computed in UInt64 and bounded
  SABRSegmentCollector  Int(id) == wantedHeaderId in collect() and
                        finishIfWanted(); now compared in the wire's own width
  MediaHeader           ticks * 1_000 overflowed a 32-bit Int; scaled in Int64
                        with overflow checked, and the fields clamped
  HLSGenerator          Int(Double(size) * 8 / duration) traps on a near-zero
                        duration that still passes the > 0 guard -- on every
                        architecture, not just 32-bit
  ChannelRSSParser      Int.init(String) is failable, so a view count above
                        2^31 silently became nil on 32-bit: the number vanished
                        for exactly the most-watched videos. Parsed as Int64.

The protobuf and UMP cases parse untrusted network data, so the first four are
reachable from a malformed response on arm64 too, and HLSGenerator's is
architecture-independent.

+++ Opaline/Core/Services/ChannelRSSParser.swift
@@ -23,7 +23,7 @@ private final class FeedDelegate: NSObject, XMLParserDelegate {
     private var videoId = ""
     private var published = ""
     private var title = ""
-    private var viewCount: Int?
+    private var viewCount: Int64?
     private let dateFormatter = ISO8601DateFormatter()
 
     func parser(
@@ -41,7 +41,10 @@ private final class FeedDelegate: NSObject, XMLParserDelegate {
             viewCount = nil
         }
         if insideEntry, elementName == "media:statistics" {
-            viewCount = (attributeDict["views"]).flatMap(Int.init)
+            // Int.init(String) is failable, so on a 32-bit target a count
+            // above 2^31 silently became nil -- the number vanished for
+            // exactly the most-viewed videos. Parse as Int64.
+            viewCount = (attributeDict["views"]).flatMap(Int64.init)
         }
         currentElement = elementName
     }
